Working on some UI stuff

+++ app/src/main/java/com/example/tradingcards/MainReceiver.kt
@@ -3,7 +3,7 @@ package com.example.tradingcards
 import com.example.tradingcards.db.DBManager
 
 interface MainReceiver {
-    fun getScreenDims(): HashMap<String, Int>
+    fun getScreenDims(): HashMap<String, Float>
     fun getDefaultDesign(width: Int, height: Int): ArrayList<HashMap<String, Any?>>
     fun getDBManager(): DBManager
 }
